params.js: slightly optimize parse for a common case
This is often reached by PopupMenu.addActor(actor) calls without
the second "params" parameter. We don't need to do two loops if we
have no params to parse. This saves a few milliseconds when large
numbers of actors are added to menus such as on startup. For individual
Params.parse() calls, the difference is in the microseconds.