window-list: simplify window previews
This creates a Clutter.Clone on show and keeps it for the lifetime of the
window preview instead of creating a new one every time it shows. We set
the size on show, but don't bother connecting to size-changed. Size changes
during preview observation are unlikely, and aren't worth the added
complexity.