window-list: add delay to dnd window activation
Prevents accidental window activations while crossing the window list
during a drag and drop operation.

This test commit breaks "draggable" actor window activations. In other
words, it only supports window activations by xdnd. I can't think of a use
case for dragging a clutter actor to an application window right now, but
if we need to support that we just have to add `lastPos` to dnd.js.

The delay is hardcoded to 500ms for now... needs testing, but it feels
too brief still imo.