menu applet: rework refreshing
This reworks all refresh sources to use a single _queueRefresh function
along with a defined set of refresh type flags and an optional delay.
The actual refresh now happens during an idle callback, or a timeout
if a delay is specified.

This ensures that all refreshing happens over the same code path so
that we can optimize the width-setting code for the applications box to
ensure that we aren't repeatedly resetting the width during each
refresh event. Additionally, it may aggregate some refreshes if
many happen during before an idle, or during the delay but this is not
likely.

These changes can make a considerable difference in responsiveness
during high-frequency refresh events such as quickly scrolling through
photos with an image viewer when recent file tracking is enabled.