lookingGlass.js: only track "common" window types
The window tracker was breaking when encountering MetaWindows for
OVERRIDE_OTHER window types which had null title/wmclass/etc because
it was trying to return null in a dbus method which has a return type
signature of aa{ss}.

To fix this I've added a list of window types to skip tracking and used
the existing objectToString function to convert any object to string.

This new behavior will exclude more window types that we don't really
need to track, and ensure any null or undefined window properties will
be converted to their respective strings.

Note: Window property changes are not currently tracked, only creation
and deletion.