window-list: use a single window preview actor
Instead of having a separate uiGroup actor for each window list item
we now just use a dummy tooltip that shows and hides a shared actor.
The functionality to resize while visible has been removed as it is
sized on show, and the window size is unlikely to change while the
preview is being shown.

This also gets rid of the access after dispose warnings in the newer
cjs.