window-tracker.c: simpler window tracking
Since we have the MetaScreen::window-added signal that is emitted
for every window regardless of workspace, we can use that instead
of tracking workspaces. In addition, instead of tracking the win-
dow removed signal, we can just connect to MetaWindow::unmanaged
on windows that we're actually keeping track of.

This also helps ensure that the window<->app mapping is initialized
by the time any of the applet signal handlers are reached, as
long as the MetaScreen::window-monitor-changed signal is not
instantly emitted when a window is created.

That emission shouldn't even be necessary as the monitor id is
provided with the MetaScreen::window-added emission.