main.js: stop using GenericContainer for uiGroup
This is based on gnome shell commits:
6b924c00c5c171ce11663c855bad2ec3d230f158
f4682748faf466498d3c73961a60099d48dc2254

This seems to reduce CPU usage while constantly dragging a window
around from 10% to 7% when nothing else is going on. The allocate
code now runs entirely in C instead of having to call into JS code,
at least for this actor.

The skip_paint replacement is a hack, but it should get the job
done and most of the time it's unused.