cinnamon-tray-manager: support loading multiple times
The stage is manages the first time, and any subsequent calls just
update the theme widget that is used to set the icon colors.

Previously after one systray removal, theme widget changes would
no longer be tracked, and an assertion failure would be triggered
as the stage was already managed.