notification applet: support clearing tray notifs by source
so you can clear specific applications' notifications, as an
alternative to clear all. Note: this doesn't affect notifications
waiting to be presented on screen, but can be setup to work that
way (excepting app-less notifications which are all grouped under
"unknown applications" (e.g. notify-send)

With this change, we also no longer clear all notifications
for an app when its window is focused or when a user clicks
one-of-many notifications from said app.